+++ src/main/kotlin/lotto/domain/model/LottoTicket.kt
@@ -7,9 +7,6 @@ class LottoTicket(
         require(numbers.size == numbers.toSet().size) {
             ERROR_LOTTO_DUPLICATE
         }
-//        require(!numbers.contains(bonusNumber)) {
-//            ERROR_BONUS_DUPLICATE
-//        }
     }
 
     fun getSize() = numbers.size
@@ -22,6 +19,5 @@ class LottoTicket(
 
     companion object {
         private const val ERROR_LOTTO_DUPLICATE = "로또 번호는 서로 중복되면 안 됩니다."
-//        private const val ERROR_BONUS_DUPLICATE = "보너스 번호는 당첨 번호와 중복되면 안 됩니다."
     }
 }
