+++ src/main/kotlin/lotto/domain/service/LottoScanner.kt
@@ -1,26 +1,14 @@
 package lotto.domain.service
 
-import lotto.domain.model.LottoNumber
 import lotto.domain.model.LottoTicket
-import lotto.domain.model.Rank
+import lotto.domain.model.WinningLotto
 
 class LottoScanner(
     private val lottoTickets: List<LottoTicket>,
-    private val winningNumbers: Set<LottoNumber>,
-    private val bonusNumber: LottoNumber,
+    private val winningLotto: WinningLotto,
 ) {
     fun getResult(): LottoResult {
-        val result = lottoTickets.map { getRank(it) }
+        val result = lottoTickets.map { winningLotto.getRank(it) }
         return LottoResult(result)
     }
-
-    fun getRank(lottoTicket: LottoTicket): Rank {
-        val countOfMatch: Int = getCountOfMatch(lottoTicket)
-        val hasBonusNumber: Boolean = hasBonusNumber(lottoTicket)
-        return Rank.valueOf(countOfMatch, hasBonusNumber)
-    }
-
-    private fun getCountOfMatch(lottoTicket: LottoTicket): Int = lottoTicket.countMatchingNumbers(winningNumbers)
-
-    private fun hasBonusNumber(lottoTicket: LottoTicket): Boolean = lottoTicket.hasBonusNumber(bonusNumber)
 }
